Add custom RPC settings
